OLrelease v6.12.0-205.92.4

+++ arch/arm64/kernel/traps.c
@@ -966,7 +966,17 @@ bool arm64_is_fatal_ras_serror(struct pt_regs *regs, unsigned long esr)
 	}
 }
 
-int __weak platform_serror(struct pt_regs *regs, unsigned int esr)
+/*
+ * Platform SError hook.
+ *
+ * This is called from the arm64 SError exception path. Although do_serror()
+ * is not marked noinstr, callers must treat this as NMI/error context:
+ * do not sleep, allocate memory, or take ordinary blocking locks.
+ *
+ * Return nonzero only for platform errors that are known to be containable
+ * and intentionally handled by the platform.
+ */
+int __weak platform_serror(struct pt_regs *regs, unsigned long esr)
 {
 	return 0;
 }
@@ -974,7 +984,7 @@ int __weak platform_serror(struct pt_regs *regs, unsigned int esr)
 void do_serror(struct pt_regs *regs, unsigned long esr)
 {
 	if (platform_serror(regs, esr))
-		return;
+		return;	/* platform handled the error */
 
 	/* non-RAS errors are not containable */
 	if (!arm64_is_ras_serror(esr) || arm64_is_fatal_ras_serror(regs, esr))
